+++ assets/javascripts/kitten/components/form/text-input-with-icon/index.js
@@ -6,8 +6,6 @@ import PropTypes from 'prop-types'
 import COLORS from '../../../constants/colors-config'
 import { VisuallyHidden } from '../../accessibility/visually-hidden'
 import classNames from 'classnames'
-import isEmpty from 'lodash/fp/isEmpty'
-import { ScreenConfig } from '../../../constants/screen-config'
 
 const StyledTextInputWithIcon = styled.div`
   position: relative;
@@ -51,27 +49,6 @@ const StyledTextInputWithIcon = styled.div`
       right: 0;
     }
   }
-
-  .k-Form-TextInputWithIcon__button {
-    border-radius: var(--input-button-border-radius);
-  }
-
-  &.k-Form-TextInput--orion {
-    --input-button-border-radius: ${pxToRem(4)};
-
-    &.k-Form-TextInput--big,
-    &.k-Form-TextInput--huge,
-    &.k-Form-TextInput--giant {
-      --input-button-border-radius: ${pxToRem(6)};
-
-      @media (min-width: ${ScreenConfig.M.min}px) {
-        --input-button-border-radius: ${pxToRem(8)};
-      }
-    }
-  }
-  &.k-Form-TextInput--rounded {
-    --input-button-border-radius: var(--input-height)
-  }
 `
 
 export const TextInputWithIcon = ({
@@ -80,7 +57,6 @@ export const TextInputWithIcon = ({
   iconPosition,
   accessibilityLabel,
   id,
-  buttonProps,
   ...others
 }) => {
   return (
@@ -105,27 +81,14 @@ export const TextInputWithIcon = ({
         )}
         disabled={disabled}
       />
-      {isEmpty(buttonProps) ? (
-        <span
-          aria-hidden="true"
-          className={classNames('k-Form-TextInputWithIcon__icon', {
-            'k-Form-TextInputWithIcon__icon--disabled': disabled,
-          })}
-        >
-          {icon}
-        </span>
-      ) : (
-        <button
-          {...buttonProps}
-          className={classNames('k-u-reset-button',
-            'k-Form-TextInputWithIcon__button',
-            'k-Form-TextInputWithIcon__icon', {
-            'k-Form-TextInputWithIcon__icon--disabled': disabled,
-          })}
-        >
-          {icon}
-        </button>
-      )}
+      <span
+        aria-hidden="true"
+        className={classNames('k-Form-TextInputWithIcon__icon', {
+          'k-Form-TextInputWithIcon__icon--disabled': disabled,
+        })}
+      >
+        {icon}
+      </span>
     </StyledTextInputWithIcon>
   )
 }
@@ -135,12 +98,10 @@ TextInputWithIcon.propTypes = {
   accessibilityLabel: PropTypes.string,
   icon: PropTypes.node.isRequired,
   iconPosition: PropTypes.oneOf(['left', 'right']),
-  buttonProps: PropTypes.object,
 }
 
 TextInputWithIcon.defaultProps = {
   accessibilityLabel: '',
   disabled: false,
   iconPosition: 'left',
-  buttonProps: {},
 }